% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterposter.R
\name{poster_better}
\alias{poster_better}
\title{Create a Better Poster in HTML}
\usage{
poster_better(..., css = NULL, hero_background = NULL,
  hero_color = NULL, qrcode = NULL, logo = NULL,
  accent_color = NULL, mathjax = FALSE, pandoc_args = NULL)
}
\value{
An R Markdown output format.
}
\description{
Create a Better Poster in HTML
}
